Make NTP request more robust.
Previously, only the memory for the NTP header was declared and set.
Actually, apart from the headers, the rest of the NTP packet (48 bytes)
contains timestamps.
Those timestamps are interpreted by the NTP server as offsets from their time
and are used to calculate the new time. Therefore, if the rest of the
packet is not zero, a wrong time is received.

By reserving memory only of the header and sending 48 bytes nonetheless
we send as timestams whatever values the microcontroller has on the stack. If
the stack content is not zero, we receive back a garbage timestamp.

This commit fixes this issue and also make the code a bit more readable
by moving some magic constants in the header file.